Migrate from --register-unschedulable+taint-and-uncordon to --register-with-taints Ref #543
