+++ core/controlplane/config/templates/cloud-config-controller
@@ -198,7 +198,7 @@ coreos:
         --rkt-path=/usr/bin/rkt \
         --rkt-stage1-image=coreos.com/rkt/stage1-coreos \
         {{if .Experimental.NodeLabels.Enabled}}--node-labels {{.Experimental.NodeLabels.String}} \
-        {{end}}--register-schedulable=false \
+        {{end}}--register-with-taints=node.alpha.kubernetes.io/role=master:NoSchedule \
         --allow-privileged=true \
         --pod-manifest-path=/etc/kubernetes/manifests \
         --cluster_dns={{.DNSServiceIP}} \
@@ -276,24 +276,6 @@ coreos:
         [Install]
         WantedBy=kubelet.service
 {{ end }}
-    - name: kube-node-taint-and-uncordon.service
-      command: start
-      runtime: true
-      content: |
-        [Unit]
-        Description=Taint this kubernetes node with its role and then uncordon it
-        Wants=kubelet.service
-        After=kubelet.service
-        Before=cfn-signal.service
-
-        [Service]
-        Type=simple
-        StartLimitInterval=0
-        RestartSec=10
-        Restart=on-failure
-        ExecStartPre=/usr/bin/systemctl is-active kubelet.service
-        ExecStartPre=/usr/bin/bash -c "while sleep 1; do if /usr/bin/curl -s -m 20 -f  http://127.0.0.1:8080/healthz > /dev/null &&  /usr/bin/curl -s -m 20 -f  http://127.0.0.1:10252/healthz > /dev/null && /usr/bin/curl -s -m 20 -f  http://127.0.0.1:10251/healthz > /dev/null &&  /usr/bin/curl --insecure -s -m 20 -f  https://127.0.0.1:10250/healthz > /dev/null ; then break ; fi;  done"
-        ExecStart=/opt/bin/taint-and-uncordon
 {{if .WaitSignal.Enabled}}
     - name: cfn-signal.service
       command: start
@@ -391,8 +373,9 @@ coreos:
       content: |
         [Unit]
         Description=Load efs persistent volume mount
-        Wants=kube-node-taint-and-uncordon.service
-        After=kube-node-taint-and-uncordon.service
+        Wants=kubelet.service
+        After=kubelet.service
+        Before=cfn-signal.service
         [Service]
         Type=simple
         RemainAfterExit=true
@@ -1010,29 +993,6 @@ write_files:
 
       rkt rm --uuid-file=/var/run/coreos/decrypt-assets.uuid || :
 {{ end }}
-
-  - path: /opt/bin/taint-and-uncordon
-    owner: root:root
-    permissions: 0700
-    content: |
-      #!/bin/bash -e
-
-      hostname=$(hostname)
-
-      docker run --rm --net=host \
-        -v /etc/kubernetes:/etc/kubernetes \
-        -v /etc/resolv.conf:/etc/resolv.conf \
-        {{.HyperkubeImage.RepoWithTag}} /bin/bash \
-          -vxec \
-          'echo "tainting this node."
-           hostname="'${hostname}'"
-           kubectl="/kubectl --server=http://127.0.0.1:8080"
-           taint="$kubectl taint node --overwrite"
-           $taint "$hostname" "node.alpha.kubernetes.io/role=master:NoSchedule"
-           echo "done."
-           echo "uncordoning this node."
-           $kubectl uncordon "$hostname"
-           echo "done."'
 
   {{if .Experimental.Plugins.Rbac.Enabled }}
   # Creates the cluster-admin role, useful for Kubernetes < 1.5.x clusters
